rename tracer to base_tracer
