+++ airflow-core/src/airflow/executors/base_executor.py
@@ -34,7 +34,7 @@ from airflow.executors.executor_loader import ExecutorLoader
 from airflow.models import Log
 from airflow.observability.stats import Stats
 from airflow.observability.traces import NO_TRACE_ID
-from airflow.observability.traces.tracer import DebugTrace, Trace, add_debug_span, gen_context
+from airflow.observability.traces.base_tracer import DebugTrace, Trace, add_debug_span, gen_context
 from airflow.observability.traces.utils import gen_span_id_from_ti_key
 from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.utils.state import TaskInstanceState
