+++ airflow-core/src/airflow/models/dagrun.py
@@ -69,7 +69,7 @@ from airflow.models.taskinstancehistory import TaskInstanceHistory as TIH
 from airflow.models.tasklog import LogTemplate
 from airflow.models.taskmap import TaskMap
 from airflow.observability.stats import Stats
-from airflow.observability.traces.tracer import EmptySpan, Trace
+from airflow.observability.traces.base_tracer import EmptySpan, Trace
 from airflow.sdk.definitions.deadline import DeadlineReference
 from airflow.serialization.definitions.notset import NOTSET, ArgNotSet, is_arg_set
 from airflow.ti_deps.dep_context import DepContext
