+++ airflow-core/tests/integration/otel/dags/otel_test_dag.py
@@ -23,7 +23,7 @@ from opentelemetry import trace
 
 from airflow import DAG
 from airflow.observability.traces import otel_tracer
-from airflow.observability.traces.tracer import Trace
+from airflow.observability.traces.base_tracer import Trace
 from airflow.sdk import chain, task
 
 logger = logging.getLogger("airflow.otel_test_dag")
