+++ airflow-core/tests/integration/otel/dags/otel_test_dag_with_pause_in_task.py
@@ -27,7 +27,7 @@ from sqlalchemy import select
 from airflow import DAG
 from airflow.models import TaskInstance
 from airflow.observability.traces import otel_tracer
-from airflow.observability.traces.tracer import Trace
+from airflow.observability.traces.base_tracer import Trace
 from airflow.providers.standard.version_compat import AIRFLOW_V_3_0_PLUS
 from airflow.sdk import chain, task
 from airflow.utils.session import create_session
